+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH19ZYN31X51PM6PPA1J1NSG"
+        "01KH2G7KVQZEH0N2714TZ8WH6A"
     )
     r"""The unique identifier of the data source"""