+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasetsop.py
@@ -113,7 +113,7 @@ class ListDatasetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-09T13:38:12.101Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-10T00:46:29.786Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")