+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH19ZYMZ2K7WDM37B0KMN491"
+        "01KH2G7KVK38N0RN8CBBXRH233"
     )
     r"""The unique identifier of the data source"""