+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatasourceop.py
@@ -90,7 +90,7 @@ class RetrieveDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KH19ZYN1H1K0EAWMJWSGV48C"
+        "01KH2G7KVNEZ64CXYZS3MWWXXV"
     )
     r"""The unique identifier of the data source"""