+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -231,7 +231,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kh19zy48pf5dnybnaenvt5e4"
+        "conv_01kh2g7kbgr1s491e5ds4xfehd"
     )
     r"""Unique conversation identifier with `conv_` prefix."""